% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_marketing_tool.r
\name{gs_marketing_tool}
\alias{gs_marketing_tool}
\title{Get Data from Steam Marketing Tool}
\usage{
gs_marketing_tool(tags, page, vr = "all")
}
\arguments{
\item{tags}{Steam tags vector}

\item{page}{Page number}

\item{vr}{VR filter}
}
\description{
Get Data from Steam Marketing Tool <https://games-stats.com/steam>
}
