% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_marketing_tool.r
\name{gs_marketing_tool}
\alias{gs_marketing_tool}
\title{Get Data from Steam Marketing Tool}
\usage{
gs_marketing_tool(tag, page)
}
\arguments{
\item{tag}{Steam tag}

\item{page}{Page number}
}
\description{
Get Data from Steam Marketing Tool <https://games-stats.com/steam>
}
